requirements.txt
nose2==0.5.0
parse==1.6.6
PyYAML==3.11
six==1.10.0
stashpy==0.1
tornado==4.2.1
tornadoes==2.4.1
wheel==0.24.0
